+++ src/components/ui/pagination.jsx
@@ -1,8 +1,8 @@
 import * as React from "react"
 import { ChevronLeft, ChevronRight, MoreHorizontal } from "lucide-react"
 
-import { cn } from "../lib/utils"
-import { buttonVariants } from "./Button2";
+import { cn } from "../../lib/utils"
+import { buttonVariants } from "./button";
 
 const Pagination = ({
   className,
