+++ src/components/ui/table.jsx
@@ -1,6 +1,6 @@
 import * as React from "react"
 
-import { cn } from "../lib/utils"
+import { cn } from "../../lib/utils"
 
 const Table = React.forwardRef(({ className, ...props }, ref) => (
   <div className="relative w-full overflow-auto shadow-xl">
